Divide header and source files into physical paths (directory structures), and logical paths (namespaces). This improves the organization of the project such that it will be easier to add more stages besides transformation.
Add getter methods to CppSig classes (macro forest related classes) for obtaining references to array members.

TODO: Check for memory leaks, rename variables related to Cpp2C to only relate to the Cpp2C transformation step
