+++ implementation/include/Transformer/TransformerSettings.hh
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace Transformer
+{
+    struct TransformerSettings
+    {
+        bool OverwriteFiles = false;
+        bool Verbose = false;
+        bool UnifyMacrosWithSameSignature = false;
+        bool OnlyCollectNotDefinedInStdHeaders = true;
+    };
+} // namespace Transformer
